+++ etc/mepo-path.py
@@ -1,3 +1,0 @@
-import os, mepo
-
-print(os.path.dirname(mepo.__file__))