+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Hidden option `--location` for mepo that returns the path to the mepo directory
+
 ### Changed
 
 ## [2.1.0] - 2024-10-02