+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Removed legacy `bin` directory
+
 ## [2.1.0] - 2024-10-02
 
 ### Fixed