+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Hidden option `--location` for mepo that returns the path to the mepo directory
+- Added ability to print number of stashes in `mepo status`
 
 ### Changed